Integer -> Long (versioned)

+++ fsl-server/src/main/java/lsfusion/erp/utils/GetActiveJavaThreadsActionProperty.java
@@ -61,7 +61,7 @@ public class GetActiveJavaThreadsActionProperty extends ScriptingActionProperty 
 
         Map<Long, Thread> threadMap = ThreadUtils.getThreadMap();
         for (ThreadInfo threadInfo : threadsInfo) {
-            int id = (int) threadInfo.getThreadId();
+            long id = threadInfo.getThreadId();
             Thread thread = threadMap.get(id);
             DataObject currentObject = new DataObject(id);
